login

+++ src/main/java/com/kevinsignes/gestion_citas/repository/IUserRepository.java
@@ -4,7 +4,7 @@ import com.kevinsignes.gestion_citas.basededatos.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface IUserRepository extends JpaRepository<UserEntity, Integer> {
+public interface IUserRepository extends JpaRepository<UserEntity, Long> {
 
     UserEntity findByUsuario(String usuario);;
    /* @Query("SELECT u FROM UserEntity u WHERE u.usuario = ?1")
